Module solana_geyser_plugin_interface::geyser_plugin_interface

Structs

ReplicaAccountInfo

Information about an account being updated

ReplicaAccountInfoV2

Information about an account being updated (extended with transaction signature doing this update)

ReplicaAccountInfoV3

Information about an account being updated (extended with reference to transaction doing this update)

ReplicaBlockInfo

ReplicaBlockInfoV2

Extending ReplicaBlockInfo by sending the transaction_entries_count.

ReplicaEntryInfo

ReplicaTransactionInfo

Information about a transaction

ReplicaTransactionInfoV2

Information about a transaction, including index in block

Enums

GeyserPluginError

Errors returned by plugin calls

ReplicaAccountInfoVersions

A wrapper to future-proof ReplicaAccountInfo handling. If there were a change to the structure of ReplicaAccountInfo, there would be new enum entry for the newer version, forcing plugin implementations to handle the change.

ReplicaBlockInfoVersions

ReplicaEntryInfoVersions

A wrapper to future-proof ReplicaEntryInfo handling. To make a change to the structure of ReplicaEntryInfo, add an new enum variant wrapping a newer version, which will force plugin implementations to handle the change.

ReplicaTransactionInfoVersions

A wrapper to future-proof ReplicaTransactionInfo handling. If there were a change to the structure of ReplicaTransactionInfo, there would be new enum entry for the newer version, forcing plugin implementations to handle the change.

SlotStatus

The current status of a slot

Traits

GeyserPlugin

Defines a Geyser plugin, to stream data from the runtime. Geyser plugins must describe desired behavior for load and unload, as well as how they will handle streamed data.

Type Definitions

Result